move simulatedbeacon package contents under eth/catalyst
